Rename to top level settings page, move advanced and schedule top level too, put settings into filter.settings.filter.xxx container, start with scrub settings. Later this can be expanded e.g. with filter.settings.nat etc...
