--- src/opnsense/mvc/app/controllers/OPNsense/Firewall/forms/advanced.xml
@@ -1,26 +1,0 @@
-<form>
-    <field>
-        <type>header</type>
-        <label>Scrub</label>
-    </field>
-    <field>
-        <id>filter.advanced.scrub.enabled</id>
-        <label>Enable interface scrub</label>
-        <type>checkbox</type>
-        <help>Enable default interface scrubbing and MSS clamping.</help>
-    </field>
-    <field>
-        <id>filter.advanced.scrub.no_df</id>
-        <label>IP Do-Not-Fragment</label>
-        <type>checkbox</type>
-        <style>scrub_option</style>
-        <help>Clear the don't-fragment bit on reassembled IPv4 packets.</help>
-    </field>
-    <field>
-        <id>filter.advanced.scrub.random_id</id>
-        <label>IP Random ID</label>
-        <type>checkbox</type>
-        <style>scrub_option</style>
-        <help>Replace predictable IPv4 identification values with random values.</help>
-    </field>
-</form>
